VS JsDbg: Explicitly release COM interfaces
The JsDbg extension was taking down Visual Studio when terminating the
debuggee.  The root cause may be the timing of the GC releasing COM
interfaces. Per MSDN docs, it is advised to explicitly release the COM interfaces
given to the Event callback.  With this change, we now do so and the
crashes are no longer seen.

Additionally, this change resets the DiaLoader so that the symbols can
change across runs.