Merged PR 33: Add support for viewing all elements in stl vectors and lists
This change introduces a stl-helpers extension, which will host helpful extended fields/array fields/type descriptions for stl types. It starts off with array fields to view the elements in stl vectors and lists, as well as a type description for the size of a vector (since there is no member on vector that specifies this).

Subsequent changes will add support for other stl types, like map.