Merged PR 42: BlinkHelpers: Move whitespace formatting to CharacterData data type description
Moving whitespace formatting behavior out of the lowest level WTFStringImpl Text type description and into the CharacterData data type description. Because the whitespace formatting functionality wraps the text in a span, adding this call at the lowest level creates problems for higher-level type descriptions that do not expect the span to be present.