Merged PR 38: Add error messages for Chromium extensions
This change adds error messages for debugger errors to AXTree and DOMTree. To share the message structure and strings detailing common problems, a "error-messages" extension is being created. This extension is not only consumed by Chromium's AXTree and DOMTree extensions, but also by Edgehtml's BoxTree extension.

As an additional refactoring change, the blink extension is being moved under the chromium extension, so that all chromium-related extensions are under one umbrella.