Merged PR 45: Introduce the concept of synthetic modules
In non-component builds of Chromium, the same type is occasionally declared in multiple modules. To handle this, the concept of a synthetic module name is being introduced. A synthetic module name is a "fake" name that represents one or more real module names. If a type is created using a synthetic module name, when that type is queried, every real module with that synthetic name will be tried until the type is successfully found. Symmetrically, when a real module name is returned from the server, it will be converted to its synthetic name (if it has one) before being consumed by higher-level extensions. This allows extended fields/array fields/type descriptions to be written once and work across multiple modules.

To prevent excessive wrong module lookups, the collection of real module names for every synthetic name will be sorted such that the last successfully looked-up module will be the first one accessed during the next lookup. This heuristic works because it is highly likely that next lookup will be for a type in the same module.