Merged PR 46: Restore the "debugger is not broken in" indicator
While adding support for target-specific extensions, the "debugger is not broken in" indicator was broken. This happened because the serving of extensions now needs to talk to the debugger to check if a given module is loaded or not. However, if the debugger is busy, JsDbg will wait infinitely for the debugger to break in before continuing to serve extensions.

To fix this, a new API called "IsDebuggerBusy" is being added to server. This API will query WinDbg or the VS debugger to check if it is broken in or not, and is now called while serving extensions. If the debugger isn't broken in, the module loaded check will be skipped and the target-specific extension will be served unconditionally. (This is okay because the user will not be able to use the extension while the debugger is running anyways.)