Merged PR 50: Support cross-module interactions with the Chromium content module
Cross-module interactions with Chromium component builds are handled by using a synthetic module name per target process to encompass multiple (actual) module names. For modules, like content, that can be seen when debugging different processes, modular equivalency needs to be established based on the process that is actively being debugged. This change adds support for this in JsDbg, with one missing piece. Currently, there is no way to detect the target process in the JsDbg client or server. This support should come in a subsequent change. Until then, this is being worked around by each Chromium extension explicitly declaring the name of the process that it targets.