Merged PR 52: Update JsDbg in response to process changes in the debugger
When the debugger is attached to multiple processes, JsDbg only shows data associated with the process that was active when JsDbg was launched. If the process is subsequently switched in the debugger, JsDbg does not respond to the update and shows stale information. With this change, JsDbg will respond to process changes in the same way it responds to other debugger changes (ex. stepping). This feature is supported with both WinDbg and the VS debugger.