Merged PR 54: Enable process and thread switching from JsDbg
This change introduces a feature that allows the active thread or process in the debugger to be changed from the JsDbg client.

In the client, two select dropdowns are being added to the JsDbg toolbar: one for the attached processes and one for the threads of the active process. To power these dropdowns, a few APIs are being added to the transport extension.
    - GetAttachedProcesses: Gets all the processes the debugger is attached to. Used to populate the processes dropdown.
    - GetTargetProcess: Gets the process that is currently being debugged. This process will be selected in the dropdown.
    - SetTargetProcess: Sets the active process in the debugger when the user selects a different process in the dropdown.
    - GetCurrentProcessThreads: Gets all the threads for the target process. Used the populate the threads dropdown.
    - GetTargetThread: Gets the thread that is currently being debugged. This thread will be selected in the dropdown.
    - SetTargetThread: Sets the active thread in the debugger when the user selects a different thread in the dropdown.

In the server, functionality is being added to power the above-mentioned APIs. For WinDbg, the target process and the list of attached processes can be queried from the system objects data member. To set the target process, we simply issue the process switch command to the debugger via the debug control object. For the Visual Studio debugger, the target process can be retrieved from the process COM object that passed in with the Event function. The attached processes list is created and maintained by listening for the process create and process destroy events. To set the active process, we use to "Debug.SetCurrentProcess" command. The server gets/sets thread information in a similar manner to process information, except that the list of threads is retrieved from the current process rather than querying the debugger.