Merged PR 56: Move persistent store and user feedback from iefs to Azure Blob Storage
This is a first in a series of changes to decouple JsDbg from iefs. This decouple is prerequisite for expanding JsDbg's reach to users outside Microsoft.

A storage account is being created for JsDbg, and the persistent store and user feedback directories have been moved into their own (separate) Azure blob containers. All existing user data has been copied from iefs into these container, so that previous data is not lost.

As an aside, the .NET version for this project is being upgraded from 4.5.2 to 4.7.2. This upgrade allows us to use the latest Azure storage APIs, rather than older versions.