Merged PR 58: ExtensionDLL: Update WinDbg extension to retrieve JsDbg extension from Azure
Currently, the JsDbg WinDbg extension simply points to a script on iefs that retrieves the necessary files (also stored on iefs) and runs them. To decouple from iefs, the scripts and launch files are being moved into Azure, and the WinDbg extension is being modified to download these locally, using wininet.

To make this work, the JsDbg scripts also need to be updated to use the local files. This has been done for the unstable script. The same will be done for the stable script once the unstable version has been validated.