Merged PR 59: JsDbg.WinDbg: Wait for debugger to break in if querying system objects throws an exception
This change fixes a bug that caused JsDbg to shutdown (disconnect from debugger) when stepping through WinDbg. This bugs occurs when the users steps quickly in WinDbg, causing certain debug system object queries from WinDbg to fail because the debugger is busy. The fix here is retry the system object query once the debugger is no longer busy. To implement this fix, some additional plumming was needed through the rest of the JsDbg server code.