Merged PR 60: Remove user feedback feature
With this change, we are removing the user feedback feature. As we move JsDbg to be open-source, there are governances that need to be followed when it comes to data collection. To avoid that overhead, we are removing features that send data about user activity. A future change will restructure the persistant storage feature to avoid sending data upstream.