[gdb] Set Displacement to 0 if it is 2*pointer size (#33)
* [gdb] Set Displacement to 0 if it is 2*pointer size

This accounts for the 2 words reserved for RTTI information. Because
the RTTI information is stored "before" the object itself, the vtable
pointer actually points to vtable + 2 pointers. See the link
in this commit for some more context.

DbgObject.symbol (and therefore DbgObject.vcast) requires displacement
to be 0 if you're logically at the beginning of the symbol.

* [gdb] Correctly initialize IsPointer64Bit

* [gdb] Better way to set IsPointer64Bit