[gdb] Resolve two TODOs
I've convinced myself that always subtracting 2*pointer size is the right
thing, because GDB always generates that offset.

Also, there is no need to wait for a breakpoint because many commands
work even while the program is running. Even if we decide to add
such queuing, it would be better to implement it server-side because
C# has better support for that kind of asynchronicity.