[gdb] Explicitly call dotnet restore before publish
This allows us to pass flags to restore, which is useful for
the .deb build process to limit the package search to
preexisting packages so we don't need to download something
from the internet and thus get more reproducible builds.

This requires using netcoreapp2.2 as the target framework,
because that's what the dotnet SDK 2.2 comes with, and
removing TargetLatestRuntimePatch.