[gdb] Simplify .deb build procedure
If we configure this as a native package, we don't need to use
git-buildpackage, which means we don't need to worry about creating git tags.

A "native" package means that the debian/ directory is part of the original
source, which is what we have. ("Normal" debian packages have the debian
packaging scripts separately from the upstream source code)

Also mention the dejagnu build requirement.