[gdb] Better handle typedefs to anonymous unions
When compiled with clang, for a typedef to an anonymous union, gdb
will not be able to find a name and so 'name' is None. When that
happens, we should just not call strip_typedefs to keep a name
to refer to the type.

Because this only happens with clang, I have not written a testcase
for this bug. I did file a clang bug (see link in comment).

This shows up with base::internal::LockImpl::NativeHandle.