Don't repeat the jsdbg command when the user presses enter
Usually, when pressing enter, GDB will repeat the last command.
This can be very useful for commands like "next". But for "jsdbg",
this is not ideal, especially when the user presses enter to get
back to the GDB prompt after the text from jsdbg has been printed.

This patch suppresses that behavior.